refactor: extract event logic into separate service layer
